Refactor and extract modal components for clarity
Extracted LayoutUsageIndicator and WorkflowHoverCard into separate files for better modularity and maintainability. Refactored EnhancedPublishModal to use new subcomponents for header, content, actions, and resource status badge, improving readability and separation of concerns.
